refactor: Split archival mutator set into its own crate
