+++ neptune-archive/src/archival_state.rs
@@ -9,6 +9,7 @@ use anyhow::bail;
 use anyhow::ensure;
 use itertools::Itertools;
 use memmap2::MmapOptions;
+use neptune_archival_mutator_set::rusty_archival_mutator_set::RustyArchivalMutatorSet;
 use neptune_consensus::block::Block;
 use neptune_consensus::block::INITIAL_BLOCK_SUBSIDY;
 use neptune_consensus::block::PREMINE_MAX_SIZE;
@@ -34,7 +35,6 @@ use neptune_mutator_set::commit;
 use neptune_mutator_set::mutator_set_accumulator::MutatorSetAccumulator;
 use neptune_mutator_set::removal_record::RemovalRecord;
 use neptune_mutator_set::removal_record::absolute_index_set::AbsoluteIndexSet;
-use neptune_mutator_set::rusty_archival_mutator_set::RustyArchivalMutatorSet;
 use neptune_primitives::block_height::BLOCKS_PER_GENERATION;
 use neptune_primitives::block_height::BlockHeight;
 use neptune_primitives::block_height::NUM_BLOCKS_SKIPPED_BECAUSE_REBOOT;
