+++ neptune-mutator-set/Cargo.toml
@@ -24,8 +24,6 @@ arbitrary-impls = [
 test-helpers = ["arbitrary-impls", "dep:rand"]
 
 [dependencies]
-neptune-archival-mmr = { version = "0.14.0", path = "../neptune-archival-mmr" }
-neptune-database = { version = "0.14.0", path = "../neptune-database" }
 get-size2 = { version = "0.7", features = ["derive"] }
 itertools = "0.14"
 num-traits = "0.2"
